add FastHealthyMealTestFactory object for reusable test data and default meal creation
